refactor: move MockFileSystem to fstest subpackage
MockFileSystem and MockFullyFeaturedFileSystem shipped in the root fs
package with no build tag, so at v1.0 they would join the public-API
compatibility promise and pollute the fs.* namespace.

Moved both types to a new github.com/ungerik/go-fs/fstest package (mirroring
testing/fstest), with all references fs.-qualified and the stale
MockFileSystemFullyFeatured panic message names corrected. The two
mock-dependent tests (TestFile, TestFile_WriteAllContext_FallbackTruncates)
moved to an external package fs_test file (file_mock_test.go, dot-importing fs)
to avoid the fstest->fs import cycle that an internal package fs test would
create. TODOS.md updated.

All root-module tests pass; go vet clean.
